Crate aws_sdk_opsworks

AWS OpsWorks

Welcome to the AWS OpsWorks Stacks API Reference. This guide provides descriptions, syntax, and usage examples for AWS OpsWorks Stacks actions and data types, including common parameters and error codes.

AWS OpsWorks Stacks is an application management service that provides an integrated experience for overseeing the complete application lifecycle. For information about this product, go to the AWS OpsWorks details page.

SDKs and CLI

The most common way to use the AWS OpsWorks Stacks API is by using the AWS Command Line Interface (CLI) or by using one of the AWS SDKs to implement applications in your preferred language. For more information, see:

AWS CLI

AWS SDK for Java

AWS SDK for .NET

AWS SDK for PHP 2

AWS SDK for Ruby

AWS SDK for Node.js

AWS SDK for Python(Boto)

Endpoints

AWS OpsWorks Stacks supports the following endpoints, all HTTPS. You must connect to one of the following endpoints. Stacks can only be accessed or managed within the endpoint in which they are created.

opsworks.us-east-1.amazonaws.com

opsworks.us-east-2.amazonaws.com

opsworks.us-west-1.amazonaws.com

opsworks.us-west-2.amazonaws.com

opsworks.ca-central-1.amazonaws.com (API only; not available in the AWS console)

opsworks.eu-west-1.amazonaws.com

opsworks.eu-west-2.amazonaws.com

opsworks.eu-west-3.amazonaws.com

opsworks.eu-central-1.amazonaws.com

opsworks.ap-northeast-1.amazonaws.com

opsworks.ap-northeast-2.amazonaws.com

opsworks.ap-south-1.amazonaws.com

opsworks.ap-southeast-1.amazonaws.com

opsworks.ap-southeast-2.amazonaws.com

opsworks.sa-east-1.amazonaws.com

Chef Versions

When you call CreateStack, CloneStack, or UpdateStack we recommend you use the ConfigurationManager parameter to specify the Chef version. The recommended and default value for Linux stacks is currently 12. Windows stacks use Chef 12.2. For more information, see Chef Versions.

You can specify Chef 12, 11.10, or 11.4 for your Linux stack. We recommend migrating your existing Linux stacks to Chef 12 as soon as possible.

Re-exports

pub use config::Config;

pub use client::Client;

Modules

client

config

error

input

model

operation

output

Structs

Blob

ByteStream

Stream of binary data

Credentials

AWS SDK Credentials

Endpoint

API Endpoint

Region

The region to send requests to.

Enums

Error

SdkError

Failed SDK Result

Statics

PKG_VERSION