Struct aws_sdk_opsworks::client::fluent_builders::UpdateInstance

pub struct UpdateInstance<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

Implementations

impl<C, M, R> UpdateInstance<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub async fn send(
self
) -> Result<UpdateInstanceOutput, SdkError<UpdateInstanceError>> where
R::Policy: SmithyRetryPolicy<UpdateInstanceInputOperationOutputAlias, UpdateInstanceOutput, UpdateInstanceError, UpdateInstanceInputOperationRetryAlias>,

pub fn instance_id(self, inp: impl Into<String>) -> Self

The instance ID.

pub fn set_instance_id(self, input: Option<String>) -> Self

pub fn layer_ids(self, inp: impl Into<String>) -> Self

Appends an item to LayerIds.

To override the contents of this collection use set_layer_ids.

The instance's layer IDs.

pub fn set_layer_ids(self, input: Option<Vec<String>>) -> Self

pub fn instance_type(self, inp: impl Into<String>) -> Self

The instance type, such as t2.micro. For a list of supported instance types, open the stack in the console, choose Instances, and choose + Instance. The Size list contains the currently supported types. For more information, see Instance Families and Types. The parameter values that you use to specify the various types are in the API Name column of the Available Instance Types table.

pub fn set_instance_type(self, input: Option<String>) -> Self

pub fn auto_scaling_type(self, inp: AutoScalingType) -> Self

For load-based or time-based instances, the type. Windows stacks can use only time-based instances.

pub fn set_auto_scaling_type(self, input: Option<AutoScalingType>) -> Self

pub fn hostname(self, inp: impl Into<String>) -> Self

The instance host name.

pub fn set_hostname(self, input: Option<String>) -> Self

pub fn os(self, inp: impl Into<String>) -> Self

The instance's operating system, which must be set to one of the following. You cannot update an instance that is using a custom AMI.

A supported Linux operating system: An Amazon Linux version, such as Amazon Linux 2018.03, Amazon Linux 2017.09, Amazon Linux 2017.03, Amazon Linux 2016.09, Amazon Linux 2016.03, Amazon Linux 2015.09, or Amazon Linux 2015.03.

A supported Ubuntu operating system, such as Ubuntu 16.04 LTS, Ubuntu 14.04 LTS, or Ubuntu 12.04 LTS.

CentOS Linux 7

Red Hat Enterprise Linux 7

A supported Windows operating system, such as Microsoft Windows Server 2012 R2 Base, Microsoft Windows Server 2012 R2 with SQL Server Express, Microsoft Windows Server 2012 R2 with SQL Server Standard, or Microsoft Windows Server 2012 R2 with SQL Server Web.

For more information about supported operating systems, see AWS OpsWorks Stacks Operating Systems.

The default option is the current Amazon Linux version. If you set this parameter to Custom, you must use the AmiId parameter to specify the custom AMI that you want to use. For more information about supported operating systems, see Operating Systems. For more information about how to use custom AMIs with OpsWorks, see Using Custom AMIs.

You can specify a different Linux operating system for the updated stack, but you cannot change from Linux to Windows or Windows to Linux.

pub fn set_os(self, input: Option<String>) -> Self

pub fn ami_id(self, inp: impl Into<String>) -> Self

The ID of the AMI that was used to create the instance. The value of this parameter must be the same AMI ID that the instance is already using. You cannot apply a new AMI to an instance by running UpdateInstance. UpdateInstance does not work on instances that are using custom AMIs.

pub fn set_ami_id(self, input: Option<String>) -> Self

pub fn ssh_key_name(self, inp: impl Into<String>) -> Self

The instance's Amazon EC2 key name.

pub fn set_ssh_key_name(self, input: Option<String>) -> Self

pub fn architecture(self, inp: Architecture) -> Self

The instance architecture. Instance types do not necessarily support both architectures. For a list of the architectures that are supported by the different instance types, see Instance Families and Types.

pub fn set_architecture(self, input: Option<Architecture>) -> Self

pub fn install_updates_on_boot(self, inp: bool) -> Self

Whether to install operating system and package updates when the instance boots. The default value is true. To control when updates are installed, set this value to false. You must then update your instances manually by using CreateDeployment to run the update_dependencies stack command or by manually running yum (Amazon Linux) or apt-get (Ubuntu) on the instances.

We strongly recommend using the default value of true, to ensure that your instances have the latest security updates.

pub fn set_install_updates_on_boot(self, input: Option<bool>) -> Self

pub fn ebs_optimized(self, inp: bool) -> Self

This property cannot be updated.

pub fn set_ebs_optimized(self, input: Option<bool>) -> Self

pub fn agent_version(self, inp: impl Into<String>) -> Self

The default AWS OpsWorks Stacks agent version. You have the following options:

INHERIT - Use the stack's default agent version setting.

version_number - Use the specified agent version. This value overrides the stack's default setting. To update the agent version, you must edit the instance configuration and specify a new version. AWS OpsWorks Stacks then automatically installs that version on the instance.

The default setting is INHERIT. To specify an agent version, you must use the complete version number, not the abbreviated number shown on the console. For a list of available agent version numbers, call DescribeAgentVersions.

AgentVersion cannot be set to Chef 12.2.

pub fn set_agent_version(self, input: Option<String>) -> Self