Struct aws_sdk_opsworks::client::Client

pub struct Client<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

An ergonomic service client for OpsWorks_20130218.

This client allows ergonomic access to a OpsWorks_20130218-shaped service. Each method corresponds to an endpoint defined in the service’s Smithy model, and the request and response shapes are auto-generated from that same model.

Using a Client

Once you have a client set up, you can access the service’s endpoints by calling the appropriate method on Client. Each such method returns a request builder for that endpoint, with methods for setting the various fields of the request. Once your request is complete, use the send method to send the request. send returns a future, which you then have to .await to get the service’s response.

Implementations

impl<C, M, R> Client<C, M, R>

pub fn with_config(client: Client<C, M, R>, conf: Config) -> Self

pub fn conf(&self) -> &Config

impl<C, M, R> Client<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub fn assign_instance(&self) -> AssignInstance<C, M, R>

pub fn assign_volume(&self) -> AssignVolume<C, M, R>

pub fn associate_elastic_ip(&self) -> AssociateElasticIp<C, M, R>

pub fn attach_elastic_load_balancer(&self) -> AttachElasticLoadBalancer<C, M, R>

pub fn clone_stack(&self) -> CloneStack<C, M, R>

pub fn create_app(&self) -> CreateApp<C, M, R>

pub fn create_deployment(&self) -> CreateDeployment<C, M, R>

pub fn create_instance(&self) -> CreateInstance<C, M, R>

pub fn create_layer(&self) -> CreateLayer<C, M, R>

pub fn create_stack(&self) -> CreateStack<C, M, R>

pub fn create_user_profile(&self) -> CreateUserProfile<C, M, R>

pub fn delete_app(&self) -> DeleteApp<C, M, R>

pub fn delete_instance(&self) -> DeleteInstance<C, M, R>

pub fn delete_layer(&self) -> DeleteLayer<C, M, R>

pub fn delete_stack(&self) -> DeleteStack<C, M, R>

pub fn delete_user_profile(&self) -> DeleteUserProfile<C, M, R>

pub fn deregister_ecs_cluster(&self) -> DeregisterEcsCluster<C, M, R>

pub fn deregister_elastic_ip(&self) -> DeregisterElasticIp<C, M, R>

pub fn deregister_instance(&self) -> DeregisterInstance<C, M, R>

pub fn deregister_rds_db_instance(&self) -> DeregisterRdsDbInstance<C, M, R>

pub fn deregister_volume(&self) -> DeregisterVolume<C, M, R>

pub fn describe_agent_versions(&self) -> DescribeAgentVersions<C, M, R>

pub fn describe_apps(&self) -> DescribeApps<C, M, R>

pub fn describe_commands(&self) -> DescribeCommands<C, M, R>

pub fn describe_deployments(&self) -> DescribeDeployments<C, M, R>

pub fn describe_ecs_clusters(&self) -> DescribeEcsClusters<C, M, R>

pub fn describe_elastic_ips(&self) -> DescribeElasticIps<C, M, R>

pub fn describe_elastic_load_balancers(
&self
) -> DescribeElasticLoadBalancers<C, M, R>

pub fn describe_instances(&self) -> DescribeInstances<C, M, R>

pub fn describe_layers(&self) -> DescribeLayers<C, M, R>

pub fn describe_load_based_auto_scaling(
&self
) -> DescribeLoadBasedAutoScaling<C, M, R>

pub fn describe_my_user_profile(&self) -> DescribeMyUserProfile<C, M, R>

pub fn describe_operating_systems(&self) -> DescribeOperatingSystems<C, M, R>

pub fn describe_permissions(&self) -> DescribePermissions<C, M, R>

pub fn describe_raid_arrays(&self) -> DescribeRaidArrays<C, M, R>

pub fn describe_rds_db_instances(&self) -> DescribeRdsDbInstances<C, M, R>

pub fn describe_service_errors(&self) -> DescribeServiceErrors<C, M, R>

pub fn describe_stack_provisioning_parameters(
&self
) -> DescribeStackProvisioningParameters<C, M, R>

pub fn describe_stacks(&self) -> DescribeStacks<C, M, R>

pub fn describe_stack_summary(&self) -> DescribeStackSummary<C, M, R>

pub fn describe_time_based_auto_scaling(
&self
) -> DescribeTimeBasedAutoScaling<C, M, R>

pub fn describe_user_profiles(&self) -> DescribeUserProfiles<C, M, R>

pub fn describe_volumes(&self) -> DescribeVolumes<C, M, R>

pub fn detach_elastic_load_balancer(&self) -> DetachElasticLoadBalancer<C, M, R>

pub fn disassociate_elastic_ip(&self) -> DisassociateElasticIp<C, M, R>

pub fn get_hostname_suggestion(&self) -> GetHostnameSuggestion<C, M, R>

pub fn grant_access(&self) -> GrantAccess<C, M, R>

pub fn list_tags(&self) -> ListTags<C, M, R>

pub fn reboot_instance(&self) -> RebootInstance<C, M, R>

pub fn register_ecs_cluster(&self) -> RegisterEcsCluster<C, M, R>

pub fn register_elastic_ip(&self) -> RegisterElasticIp<C, M, R>

pub fn register_instance(&self) -> RegisterInstance<C, M, R>

pub fn register_rds_db_instance(&self) -> RegisterRdsDbInstance<C, M, R>

pub fn register_volume(&self) -> RegisterVolume<C, M, R>

pub fn set_load_based_auto_scaling(&self) -> SetLoadBasedAutoScaling<C, M, R>

pub fn set_permission(&self) -> SetPermission<C, M, R>

pub fn set_time_based_auto_scaling(&self) -> SetTimeBasedAutoScaling<C, M, R>

pub fn start_instance(&self) -> StartInstance<C, M, R>

pub fn start_stack(&self) -> StartStack<C, M, R>

pub fn stop_instance(&self) -> StopInstance<C, M, R>

pub fn stop_stack(&self) -> StopStack<C, M, R>

pub fn tag_resource(&self) -> TagResource<C, M, R>

pub fn unassign_instance(&self) -> UnassignInstance<C, M, R>

pub fn unassign_volume(&self) -> UnassignVolume<C, M, R>

pub fn untag_resource(&self) -> UntagResource<C, M, R>

pub fn update_app(&self) -> UpdateApp<C, M, R>

pub fn update_elastic_ip(&self) -> UpdateElasticIp<C, M, R>

pub fn update_instance(&self) -> UpdateInstance<C, M, R>

pub fn update_layer(&self) -> UpdateLayer<C, M, R>

pub fn update_my_user_profile(&self) -> UpdateMyUserProfile<C, M, R>

pub fn update_rds_db_instance(&self) -> UpdateRdsDbInstance<C, M, R>

pub fn update_stack(&self) -> UpdateStack<C, M, R>

pub fn update_user_profile(&self) -> UpdateUserProfile<C, M, R>

pub fn update_volume(&self) -> UpdateVolume<C, M, R>

impl<C> Client<C, AwsMiddleware, Standard>

pub fn from_conf_conn(conf: Config, conn: C) -> Self

impl Client<DynConnector, AwsMiddleware, Standard>

pub fn new(config: &Config) -> Self

pub fn from_conf(conf: Config) -> Self