Module aws_sdk_opsworks::model

Modules

agent_version

See AgentVersion

app

See App

auto_scaling_thresholds

See AutoScalingThresholds

block_device_mapping

See BlockDeviceMapping

chef_configuration

See ChefConfiguration

cloud_watch_logs_configuration

See CloudWatchLogsConfiguration

cloud_watch_logs_log_stream

See CloudWatchLogsLogStream

command

See Command

data_source

See DataSource

deployment

See Deployment

deployment_command

See DeploymentCommand

ebs_block_device

See EbsBlockDevice

ecs_cluster

See EcsCluster

elastic_ip

See ElasticIp

elastic_load_balancer

See ElasticLoadBalancer

environment_variable

See EnvironmentVariable

instance

See Instance

instance_identity

See InstanceIdentity

instances_count

See InstancesCount

layer

See Layer

lifecycle_event_configuration

See LifecycleEventConfiguration

load_based_auto_scaling_configuration

See LoadBasedAutoScalingConfiguration

operating_system

See OperatingSystem

operating_system_configuration_manager

See OperatingSystemConfigurationManager

permission

See Permission

raid_array

See RaidArray

rds_db_instance

See RdsDbInstance

recipes

See Recipes

reported_os

See ReportedOs

self_user_profile

See SelfUserProfile

service_error

See ServiceError

shutdown_event_configuration

See ShutdownEventConfiguration

source

See Source

ssl_configuration

See SslConfiguration

stack

See Stack

stack_configuration_manager

See StackConfigurationManager

stack_summary

See StackSummary

temporary_credential

See TemporaryCredential

time_based_auto_scaling_configuration

See TimeBasedAutoScalingConfiguration

user_profile

See UserProfile

volume

See Volume

volume_configuration

See VolumeConfiguration

weekly_auto_scaling_schedule

See WeeklyAutoScalingSchedule

Structs

AgentVersion

Describes an agent version.

App

A description of the app.

AutoScalingThresholds

Describes a load-based auto scaling upscaling or downscaling threshold configuration, which specifies when AWS OpsWorks Stacks starts or stops load-based instances.

BlockDeviceMapping

Describes a block device mapping. This data type maps directly to the Amazon EC2 BlockDeviceMapping data type.

ChefConfiguration

Describes the Chef configuration.

CloudWatchLogsConfiguration

Describes the Amazon CloudWatch logs configuration for a layer.

CloudWatchLogsLogStream

Describes the Amazon CloudWatch logs configuration for a layer. For detailed information about members of this data type, see the CloudWatch Logs Agent Reference.

Command

Describes a command.

DataSource

Describes an app's data source.

Deployment

Describes a deployment of a stack or app.

DeploymentCommand

Used to specify a stack or deployment command.

EbsBlockDevice

Describes an Amazon EBS volume. This data type maps directly to the Amazon EC2 EbsBlockDevice

EcsCluster

Describes a registered Amazon ECS cluster.

ElasticIp

Describes an Elastic IP address.

ElasticLoadBalancer

Describes an Elastic Load Balancing instance.

EnvironmentVariable

Represents an app's environment variable.

Instance

Describes an instance.

InstanceIdentity

Contains a description of an Amazon EC2 instance from the Amazon EC2 metadata service. For more

InstancesCount

Describes how many instances a stack has for each status.

Layer

Describes a layer.

LifecycleEventConfiguration

Specifies the lifecycle event configuration

LoadBasedAutoScalingConfiguration

Describes a layer's load-based auto scaling configuration.

OperatingSystem

Describes supported operating systems in AWS OpsWorks Stacks.

OperatingSystemConfigurationManager

A block that contains information about the configuration manager (Chef) and the versions of the configuration manager that are supported for an operating system.

Permission

Describes stack or user permissions.

RaidArray

Describes an instance's RAID array.

RdsDbInstance

Describes an Amazon RDS instance.

Recipes

AWS OpsWorks Stacks supports five

ReportedOs

A registered instance's reported operating system.

SelfUserProfile

Describes a user's SSH information.

ServiceError

Describes an AWS OpsWorks Stacks service error.

ShutdownEventConfiguration

The Shutdown event configuration.

Source

Contains the information required to retrieve an app or cookbook from a repository. For more

SslConfiguration

Describes an app's SSL configuration.

Stack

Describes a stack.

StackConfigurationManager

Describes the configuration manager.

StackSummary

Summarizes the number of layers, instances, and apps in a stack.

TemporaryCredential

Contains the data needed by RDP clients such as the Microsoft Remote Desktop Connection to log in to the instance.

TimeBasedAutoScalingConfiguration

Describes an instance's time-based auto scaling configuration.

UserProfile

Describes a user's SSH information.

Volume

Describes an instance's Amazon EBS volume.

VolumeConfiguration

Describes an Amazon EBS volume configuration.

WeeklyAutoScalingSchedule

Describes a time-based instance's auto scaling schedule. The schedule consists of a set of key-value pairs.

Enums

AppAttributesKeys

AppType

Architecture

AutoScalingType

CloudWatchLogsEncoding

Specifies the encoding of the log file so that the file can be read correctly. The default is utf_8. Encodings supported by Python codecs.decode() can be used here.

CloudWatchLogsInitialPosition

Specifies where to start to read data (start_of_file or end_of_file). The default is start_of_file. It's only used if there is no state persisted for that log stream.

CloudWatchLogsTimeZone

The preferred time zone for logs streamed to CloudWatch Logs. Valid values are LOCAL and UTC, for Coordinated Universal Time.

DeploymentCommandName

LayerAttributesKeys

LayerType

RootDeviceType

SourceType

StackAttributesKeys

VirtualizationType

VolumeType