Struct aws_sdk_opsworks::model::stack::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for Stack

Implementations

impl Builder

pub fn stack_id(self, input: impl Into<String>) -> Self

The stack ID.

pub fn set_stack_id(self, input: Option<String>) -> Self

pub fn name(self, input: impl Into<String>) -> Self

The stack name.

pub fn set_name(self, input: Option<String>) -> Self

pub fn arn(self, input: impl Into<String>) -> Self

The stack's ARN.

pub fn set_arn(self, input: Option<String>) -> Self

pub fn region(self, input: impl Into<String>) -> Self

The stack AWS region, such as "ap-northeast-2". For more information about AWS regions, see Regions and Endpoints.

pub fn set_region(self, input: Option<String>) -> Self

pub fn vpc_id(self, input: impl Into<String>) -> Self

The VPC ID; applicable only if the stack is running in a VPC.

pub fn set_vpc_id(self, input: Option<String>) -> Self

pub fn attributes(
self,
k: impl Into<StackAttributesKeys>,
v: impl Into<String>
) -> Self

pub fn set_attributes(
self,
input: Option<HashMap<StackAttributesKeys, String>>
) -> Self

pub fn service_role_arn(self, input: impl Into<String>) -> Self

The stack AWS Identity and Access Management (IAM) role.

pub fn set_service_role_arn(self, input: Option<String>) -> Self

pub fn default_instance_profile_arn(self, input: impl Into<String>) -> Self

The ARN of an IAM profile that is the default profile for all of the stack's EC2 instances. For more information about IAM ARNs, see Using Identifiers.

pub fn set_default_instance_profile_arn(self, input: Option<String>) -> Self

pub fn default_os(self, input: impl Into<String>) -> Self

The stack's default operating system.

pub fn set_default_os(self, input: Option<String>) -> Self

pub fn hostname_theme(self, input: impl Into<String>) -> Self

The stack host name theme, with spaces replaced by underscores.

pub fn set_hostname_theme(self, input: Option<String>) -> Self

pub fn default_availability_zone(self, input: impl Into<String>) -> Self

The stack's default Availability Zone. For more information, see Regions and Endpoints.

pub fn set_default_availability_zone(self, input: Option<String>) -> Self

pub fn default_subnet_id(self, input: impl Into<String>) -> Self

The default subnet ID; applicable only if the stack is running in a VPC.

pub fn set_default_subnet_id(self, input: Option<String>) -> Self

pub fn custom_json(self, input: impl Into<String>) -> Self

A JSON object that contains user-defined attributes to be added to the stack configuration and deployment attributes. You can use custom JSON to override the corresponding default stack configuration attribute values or to pass data to recipes. The string should be in the following format:

"{\"key1\": \"value1\", \"key2\": \"value2\",...}"

For more information on custom JSON, see Use Custom JSON to Modify the Stack Configuration Attributes.

pub fn set_custom_json(self, input: Option<String>) -> Self

pub fn configuration_manager(self, input: StackConfigurationManager) -> Self

The configuration manager.

pub fn set_configuration_manager(
self,
input: Option<StackConfigurationManager>
) -> Self

pub fn chef_configuration(self, input: ChefConfiguration) -> Self

A ChefConfiguration object that specifies whether to enable Berkshelf and the Berkshelf version. For more information, see Create a New Stack.

pub fn set_chef_configuration(self, input: Option<ChefConfiguration>) -> Self

pub fn use_custom_cookbooks(self, input: bool) -> Self

Whether the stack uses custom cookbooks.

pub fn set_use_custom_cookbooks(self, input: Option<bool>) -> Self

pub fn use_opsworks_security_groups(self, input: bool) -> Self

Whether the stack automatically associates the AWS OpsWorks Stacks built-in security groups with the stack's layers.

pub fn set_use_opsworks_security_groups(self, input: Option<bool>) -> Self

pub fn custom_cookbooks_source(self, input: Source) -> Self

Contains the information required to retrieve an app or cookbook from a repository. For more information, see Adding Apps or Cookbooks and Recipes.

pub fn set_custom_cookbooks_source(self, input: Option<Source>) -> Self

pub fn default_ssh_key_name(self, input: impl Into<String>) -> Self

A default Amazon EC2 key pair for the stack's instances. You can override this value when you create or update an instance.

pub fn set_default_ssh_key_name(self, input: Option<String>) -> Self

pub fn created_at(self, input: impl Into<String>) -> Self

The date when the stack was created.

pub fn set_created_at(self, input: Option<String>) -> Self

pub fn default_root_device_type(self, input: RootDeviceType) -> Self

The default root device type. This value is used by default for all instances in the stack, but you can override it when you create an instance. For more information, see Storage for the Root Device.

pub fn set_default_root_device_type(self, input: Option<RootDeviceType>) -> Self

pub fn agent_version(self, input: impl Into<String>) -> Self

The agent version. This parameter is set to LATEST for auto-update. or a version number for a fixed agent version.

pub fn set_agent_version(self, input: Option<String>) -> Self

pub fn build(self) -> Stack

Consumes the builder and constructs a Stack