Module aws_sdk_opsworks::operation

Structs

AssignInstance

Assign a registered instance to a layer.

AssignVolume

Assigns one of the stack's registered Amazon EBS volumes to a specified instance. The volume must

AssociateElasticIp

Associates one of the stack's registered Elastic IP addresses with a specified instance. The

AttachElasticLoadBalancer

Attaches an Elastic Load Balancing load balancer to a specified layer. AWS OpsWorks Stacks does not support

CloneStack

Creates a clone of a specified stack. For more information, see Clone a

CreateApp

Creates an app for a specified stack. For more information, see Creating

CreateDeployment

Runs deployment or stack commands. For more information, see Deploying

CreateInstance

Creates an instance in a specified stack. For more information, see Adding an

CreateLayer

Creates a layer. For more information, see How to

CreateStack

Creates a new stack. For more information, see Create a New

CreateUserProfile

Creates a new user profile.

DeleteApp

Deletes a specified app.

DeleteInstance

Deletes a specified instance, which terminates the associated Amazon EC2 instance. You must stop an instance before you can delete it.

DeleteLayer

Deletes a specified layer. You must first stop and then delete all associated instances or

DeleteStack

Deletes a specified stack. You must first delete all instances, layers, and apps or

DeleteUserProfile

Deletes a user profile.

DeregisterEcsCluster

Deregisters a specified Amazon ECS cluster from a stack.

DeregisterElasticIp

Deregisters a specified Elastic IP address. The address can then be registered by another

DeregisterInstance

Deregister a registered Amazon EC2 or on-premises instance. This action removes the

DeregisterRdsDbInstance

Deregisters an Amazon RDS instance.

DeregisterVolume

Deregisters an Amazon EBS volume. The volume can then be registered by another stack. For more

DescribeAgentVersions

Describes the available AWS OpsWorks Stacks agent versions. You must specify a stack ID or a

DescribeApps

Requests a description of a specified set of apps.

DescribeCommands

Describes the results of specified commands.

DescribeDeployments

Requests a description of a specified set of deployments.

DescribeEcsClusters

Describes Amazon ECS clusters that are registered with a stack. If you specify only a stack ID,

DescribeElasticIps

Describes Elastic IP addresses.

DescribeElasticLoadBalancers

Describes a stack's Elastic Load Balancing instances.

DescribeInstances

Requests a description of a set of instances.

DescribeLayers

Requests a description of one or more layers in a specified stack.

DescribeLoadBasedAutoScaling

Describes load-based auto scaling configurations for specified layers.

DescribeMyUserProfile

Describes a user's SSH information.

DescribeOperatingSystems

Describes the operating systems that are supported by AWS OpsWorks Stacks.

DescribePermissions

Describes the permissions for a specified stack.

DescribeRaidArrays

Describe an instance's RAID arrays.

DescribeRdsDbInstances

Describes Amazon RDS instances.

DescribeServiceErrors

Describes AWS OpsWorks Stacks service errors.

DescribeStackProvisioningParameters

Requests a description of a stack's provisioning parameters.

DescribeStackSummary

Describes the number of layers and apps in a specified stack, and the number of instances in

DescribeStacks

Requests a description of one or more stacks.

DescribeTimeBasedAutoScaling

Describes time-based auto scaling configurations for specified instances.

DescribeUserProfiles

Describe specified users.

DescribeVolumes

Describes an instance's Amazon EBS volumes.

DetachElasticLoadBalancer

Detaches a specified Elastic Load Balancing instance from its layer.

DisassociateElasticIp

Disassociates an Elastic IP address from its instance. The address remains registered with

GetHostnameSuggestion

Gets a generated host name for the specified layer, based on the current host name theme.

GrantAccess

ListTags

Returns a list of tags that are applied to the specified stack or layer.

RebootInstance

Reboots a specified instance. For more information, see Starting,

RegisterEcsCluster

Registers a specified Amazon ECS cluster with a stack. You can register only one

RegisterElasticIp

Registers an Elastic IP address with a specified stack. An address can be registered with

RegisterInstance

Registers instances that were created outside of AWS OpsWorks Stacks with a specified stack.

RegisterRdsDbInstance

Registers an Amazon RDS instance with a stack.

RegisterVolume

Registers an Amazon EBS volume with a specified stack. A volume can be registered with only one

SetLoadBasedAutoScaling

Specify the load-based auto scaling configuration for a specified layer. For more

SetPermission

Specifies a user's permissions. For more information, see Security and

SetTimeBasedAutoScaling

Specify the time-based auto scaling configuration for a specified instance. For more

StartInstance

Starts a specified instance. For more information, see Starting,

StartStack

Starts a stack's instances.

StopInstance

Stops a specified instance. When you stop a standard instance, the data disappears and must

StopStack

Stops a specified stack.

TagResource

Apply cost-allocation tags to a specified stack or layer in AWS OpsWorks Stacks. For more information about how tagging works, see Tags in the AWS OpsWorks User Guide.

UnassignInstance

Unassigns a registered instance from all layers that are using the instance.

UnassignVolume

Unassigns an assigned Amazon EBS volume. The volume remains registered with the stack. For more

UntagResource

Removes tags from a specified stack or layer.

UpdateApp

Updates a specified app.

UpdateElasticIp

Updates a registered Elastic IP address's name. For more information, see Resource Management.

UpdateInstance

Updates a specified instance.

UpdateLayer

Updates a specified layer.

UpdateMyUserProfile

Updates a user's SSH public key.

UpdateRdsDbInstance

Updates an Amazon RDS instance.

UpdateStack

Updates a specified stack.

UpdateUserProfile

Updates a specified user profile.

UpdateVolume

Updates an Amazon EBS volume's name or mount point. For more information, see Resource Management.