Struct aws_sdk_opsworks::client::Client

pub struct Client<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

An ergonomic service client for OpsWorks_20130218.

This client allows ergonomic access to a OpsWorks_20130218-shaped service. Each method corresponds to an endpoint defined in the service’s Smithy model, and the request and response shapes are auto-generated from that same model.

Using a Client

Once you have a client set up, you can access the service’s endpoints by calling the appropriate method on Client. Each such method returns a request builder for that endpoint, with methods for setting the various fields of the request. Once your request is complete, use the send method to send the request. send returns a future, which you then have to .await to get the service’s response.

Implementations

impl<C, M, R> Client<C, M, R>

pub fn with_config(client: Client<C, M, R>, conf: Config) -> Self

Creates a client with the given service configuration.

pub fn conf(&self) -> &Config

Returns the client’s configuration.

impl<C, M, R> Client<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub fn assign_instance(&self) -> AssignInstance<C, M, R>

Constructs a fluent builder for the AssignInstance operation.

See AssignInstance for more information about the operation and its arguments.

pub fn assign_volume(&self) -> AssignVolume<C, M, R>

Constructs a fluent builder for the AssignVolume operation.

See AssignVolume for more information about the operation and its arguments.

pub fn associate_elastic_ip(&self) -> AssociateElasticIp<C, M, R>

Constructs a fluent builder for the AssociateElasticIp operation.

See AssociateElasticIp for more information about the operation and its arguments.

pub fn attach_elastic_load_balancer(&self) -> AttachElasticLoadBalancer<C, M, R>

Constructs a fluent builder for the AttachElasticLoadBalancer operation.

See AttachElasticLoadBalancer for more information about the operation and its arguments.

pub fn clone_stack(&self) -> CloneStack<C, M, R>

Constructs a fluent builder for the CloneStack operation.

See CloneStack for more information about the operation and its arguments.

pub fn create_app(&self) -> CreateApp<C, M, R>

Constructs a fluent builder for the CreateApp operation.

See CreateApp for more information about the operation and its arguments.

pub fn create_deployment(&self) -> CreateDeployment<C, M, R>

Constructs a fluent builder for the CreateDeployment operation.

See CreateDeployment for more information about the operation and its arguments.

pub fn create_instance(&self) -> CreateInstance<C, M, R>

Constructs a fluent builder for the CreateInstance operation.

See CreateInstance for more information about the operation and its arguments.

pub fn create_layer(&self) -> CreateLayer<C, M, R>

Constructs a fluent builder for the CreateLayer operation.

See CreateLayer for more information about the operation and its arguments.

pub fn create_stack(&self) -> CreateStack<C, M, R>

Constructs a fluent builder for the CreateStack operation.

See CreateStack for more information about the operation and its arguments.

pub fn create_user_profile(&self) -> CreateUserProfile<C, M, R>

Constructs a fluent builder for the CreateUserProfile operation.

See CreateUserProfile for more information about the operation and its arguments.

pub fn delete_app(&self) -> DeleteApp<C, M, R>

Constructs a fluent builder for the DeleteApp operation.

See DeleteApp for more information about the operation and its arguments.

pub fn delete_instance(&self) -> DeleteInstance<C, M, R>

Constructs a fluent builder for the DeleteInstance operation.

See DeleteInstance for more information about the operation and its arguments.

pub fn delete_layer(&self) -> DeleteLayer<C, M, R>

Constructs a fluent builder for the DeleteLayer operation.

See DeleteLayer for more information about the operation and its arguments.

pub fn delete_stack(&self) -> DeleteStack<C, M, R>

Constructs a fluent builder for the DeleteStack operation.

See DeleteStack for more information about the operation and its arguments.

pub fn delete_user_profile(&self) -> DeleteUserProfile<C, M, R>

Constructs a fluent builder for the DeleteUserProfile operation.

See DeleteUserProfile for more information about the operation and its arguments.

pub fn deregister_ecs_cluster(&self) -> DeregisterEcsCluster<C, M, R>

Constructs a fluent builder for the DeregisterEcsCluster operation.

See DeregisterEcsCluster for more information about the operation and its arguments.

pub fn deregister_elastic_ip(&self) -> DeregisterElasticIp<C, M, R>

Constructs a fluent builder for the DeregisterElasticIp operation.

See DeregisterElasticIp for more information about the operation and its arguments.

pub fn deregister_instance(&self) -> DeregisterInstance<C, M, R>

Constructs a fluent builder for the DeregisterInstance operation.

See DeregisterInstance for more information about the operation and its arguments.

pub fn deregister_rds_db_instance(&self) -> DeregisterRdsDbInstance<C, M, R>

Constructs a fluent builder for the DeregisterRdsDbInstance operation.

See DeregisterRdsDbInstance for more information about the operation and its arguments.

pub fn deregister_volume(&self) -> DeregisterVolume<C, M, R>

Constructs a fluent builder for the DeregisterVolume operation.

See DeregisterVolume for more information about the operation and its arguments.

pub fn describe_agent_versions(&self) -> DescribeAgentVersions<C, M, R>

Constructs a fluent builder for the DescribeAgentVersions operation.

See DescribeAgentVersions for more information about the operation and its arguments.

pub fn describe_apps(&self) -> DescribeApps<C, M, R>

Constructs a fluent builder for the DescribeApps operation.

See DescribeApps for more information about the operation and its arguments.

pub fn describe_commands(&self) -> DescribeCommands<C, M, R>

Constructs a fluent builder for the DescribeCommands operation.

See DescribeCommands for more information about the operation and its arguments.

pub fn describe_deployments(&self) -> DescribeDeployments<C, M, R>

Constructs a fluent builder for the DescribeDeployments operation.

See DescribeDeployments for more information about the operation and its arguments.

pub fn describe_ecs_clusters(&self) -> DescribeEcsClusters<C, M, R>

Constructs a fluent builder for the DescribeEcsClusters operation.

See DescribeEcsClusters for more information about the operation and its arguments.

pub fn describe_elastic_ips(&self) -> DescribeElasticIps<C, M, R>

Constructs a fluent builder for the DescribeElasticIps operation.

See DescribeElasticIps for more information about the operation and its arguments.

pub fn describe_elastic_load_balancers(
&self
) -> DescribeElasticLoadBalancers<C, M, R>

Constructs a fluent builder for the DescribeElasticLoadBalancers operation.

See DescribeElasticLoadBalancers for more information about the operation and its arguments.

pub fn describe_instances(&self) -> DescribeInstances<C, M, R>

Constructs a fluent builder for the DescribeInstances operation.

See DescribeInstances for more information about the operation and its arguments.

pub fn describe_layers(&self) -> DescribeLayers<C, M, R>

Constructs a fluent builder for the DescribeLayers operation.

See DescribeLayers for more information about the operation and its arguments.

pub fn describe_load_based_auto_scaling(
&self
) -> DescribeLoadBasedAutoScaling<C, M, R>

Constructs a fluent builder for the DescribeLoadBasedAutoScaling operation.

See DescribeLoadBasedAutoScaling for more information about the operation and its arguments.

pub fn describe_my_user_profile(&self) -> DescribeMyUserProfile<C, M, R>

Constructs a fluent builder for the DescribeMyUserProfile operation.

See DescribeMyUserProfile for more information about the operation and its arguments.

pub fn describe_operating_systems(&self) -> DescribeOperatingSystems<C, M, R>

Constructs a fluent builder for the DescribeOperatingSystems operation.

See DescribeOperatingSystems for more information about the operation and its arguments.

pub fn describe_permissions(&self) -> DescribePermissions<C, M, R>

Constructs a fluent builder for the DescribePermissions operation.

See DescribePermissions for more information about the operation and its arguments.

pub fn describe_raid_arrays(&self) -> DescribeRaidArrays<C, M, R>

Constructs a fluent builder for the DescribeRaidArrays operation.

See DescribeRaidArrays for more information about the operation and its arguments.

pub fn describe_rds_db_instances(&self) -> DescribeRdsDbInstances<C, M, R>

Constructs a fluent builder for the DescribeRdsDbInstances operation.

See DescribeRdsDbInstances for more information about the operation and its arguments.

pub fn describe_service_errors(&self) -> DescribeServiceErrors<C, M, R>

Constructs a fluent builder for the DescribeServiceErrors operation.

See DescribeServiceErrors for more information about the operation and its arguments.

pub fn describe_stack_provisioning_parameters(
&self
) -> DescribeStackProvisioningParameters<C, M, R>

Constructs a fluent builder for the DescribeStackProvisioningParameters operation.

See DescribeStackProvisioningParameters for more information about the operation and its arguments.

pub fn describe_stacks(&self) -> DescribeStacks<C, M, R>

Constructs a fluent builder for the DescribeStacks operation.

See DescribeStacks for more information about the operation and its arguments.

pub fn describe_stack_summary(&self) -> DescribeStackSummary<C, M, R>

Constructs a fluent builder for the DescribeStackSummary operation.

See DescribeStackSummary for more information about the operation and its arguments.

pub fn describe_time_based_auto_scaling(
&self
) -> DescribeTimeBasedAutoScaling<C, M, R>

Constructs a fluent builder for the DescribeTimeBasedAutoScaling operation.

See DescribeTimeBasedAutoScaling for more information about the operation and its arguments.

pub fn describe_user_profiles(&self) -> DescribeUserProfiles<C, M, R>

Constructs a fluent builder for the DescribeUserProfiles operation.

See DescribeUserProfiles for more information about the operation and its arguments.

pub fn describe_volumes(&self) -> DescribeVolumes<C, M, R>

Constructs a fluent builder for the DescribeVolumes operation.

See DescribeVolumes for more information about the operation and its arguments.

pub fn detach_elastic_load_balancer(&self) -> DetachElasticLoadBalancer<C, M, R>

Constructs a fluent builder for the DetachElasticLoadBalancer operation.

See DetachElasticLoadBalancer for more information about the operation and its arguments.

pub fn disassociate_elastic_ip(&self) -> DisassociateElasticIp<C, M, R>

Constructs a fluent builder for the DisassociateElasticIp operation.

See DisassociateElasticIp for more information about the operation and its arguments.

pub fn get_hostname_suggestion(&self) -> GetHostnameSuggestion<C, M, R>

Constructs a fluent builder for the GetHostnameSuggestion operation.

See GetHostnameSuggestion for more information about the operation and its arguments.

pub fn grant_access(&self) -> GrantAccess<C, M, R>

Constructs a fluent builder for the GrantAccess operation.

See GrantAccess for more information about the operation and its arguments.

pub fn list_tags(&self) -> ListTags<C, M, R>

Constructs a fluent builder for the ListTags operation.

See ListTags for more information about the operation and its arguments.

pub fn reboot_instance(&self) -> RebootInstance<C, M, R>

Constructs a fluent builder for the RebootInstance operation.

See RebootInstance for more information about the operation and its arguments.

pub fn register_ecs_cluster(&self) -> RegisterEcsCluster<C, M, R>

Constructs a fluent builder for the RegisterEcsCluster operation.

See RegisterEcsCluster for more information about the operation and its arguments.

pub fn register_elastic_ip(&self) -> RegisterElasticIp<C, M, R>

Constructs a fluent builder for the RegisterElasticIp operation.

See RegisterElasticIp for more information about the operation and its arguments.

pub fn register_instance(&self) -> RegisterInstance<C, M, R>

Constructs a fluent builder for the RegisterInstance operation.

See RegisterInstance for more information about the operation and its arguments.

pub fn register_rds_db_instance(&self) -> RegisterRdsDbInstance<C, M, R>

Constructs a fluent builder for the RegisterRdsDbInstance operation.

See RegisterRdsDbInstance for more information about the operation and its arguments.

pub fn register_volume(&self) -> RegisterVolume<C, M, R>

Constructs a fluent builder for the RegisterVolume operation.

See RegisterVolume for more information about the operation and its arguments.

pub fn set_load_based_auto_scaling(&self) -> SetLoadBasedAutoScaling<C, M, R>

Constructs a fluent builder for the SetLoadBasedAutoScaling operation.

See SetLoadBasedAutoScaling for more information about the operation and its arguments.

pub fn set_permission(&self) -> SetPermission<C, M, R>

Constructs a fluent builder for the SetPermission operation.

See SetPermission for more information about the operation and its arguments.

pub fn set_time_based_auto_scaling(&self) -> SetTimeBasedAutoScaling<C, M, R>

Constructs a fluent builder for the SetTimeBasedAutoScaling operation.

See SetTimeBasedAutoScaling for more information about the operation and its arguments.

pub fn start_instance(&self) -> StartInstance<C, M, R>

Constructs a fluent builder for the StartInstance operation.

See StartInstance for more information about the operation and its arguments.

pub fn start_stack(&self) -> StartStack<C, M, R>

Constructs a fluent builder for the StartStack operation.

See StartStack for more information about the operation and its arguments.

pub fn stop_instance(&self) -> StopInstance<C, M, R>

Constructs a fluent builder for the StopInstance operation.

See StopInstance for more information about the operation and its arguments.

pub fn stop_stack(&self) -> StopStack<C, M, R>

Constructs a fluent builder for the StopStack operation.

See StopStack for more information about the operation and its arguments.

pub fn tag_resource(&self) -> TagResource<C, M, R>

Constructs a fluent builder for the TagResource operation.

See TagResource for more information about the operation and its arguments.

pub fn unassign_instance(&self) -> UnassignInstance<C, M, R>

Constructs a fluent builder for the UnassignInstance operation.

See UnassignInstance for more information about the operation and its arguments.

pub fn unassign_volume(&self) -> UnassignVolume<C, M, R>

Constructs a fluent builder for the UnassignVolume operation.

See UnassignVolume for more information about the operation and its arguments.

pub fn untag_resource(&self) -> UntagResource<C, M, R>

Constructs a fluent builder for the UntagResource operation.

See UntagResource for more information about the operation and its arguments.

pub fn update_app(&self) -> UpdateApp<C, M, R>

Constructs a fluent builder for the UpdateApp operation.

See UpdateApp for more information about the operation and its arguments.

pub fn update_elastic_ip(&self) -> UpdateElasticIp<C, M, R>

Constructs a fluent builder for the UpdateElasticIp operation.

See UpdateElasticIp for more information about the operation and its arguments.

pub fn update_instance(&self) -> UpdateInstance<C, M, R>

Constructs a fluent builder for the UpdateInstance operation.

See UpdateInstance for more information about the operation and its arguments.

pub fn update_layer(&self) -> UpdateLayer<C, M, R>

Constructs a fluent builder for the UpdateLayer operation.

See UpdateLayer for more information about the operation and its arguments.

pub fn update_my_user_profile(&self) -> UpdateMyUserProfile<C, M, R>

Constructs a fluent builder for the UpdateMyUserProfile operation.

See UpdateMyUserProfile for more information about the operation and its arguments.

pub fn update_rds_db_instance(&self) -> UpdateRdsDbInstance<C, M, R>

Constructs a fluent builder for the UpdateRdsDbInstance operation.

See UpdateRdsDbInstance for more information about the operation and its arguments.

pub fn update_stack(&self) -> UpdateStack<C, M, R>

Constructs a fluent builder for the UpdateStack operation.

See UpdateStack for more information about the operation and its arguments.

pub fn update_user_profile(&self) -> UpdateUserProfile<C, M, R>

Constructs a fluent builder for the UpdateUserProfile operation.

See UpdateUserProfile for more information about the operation and its arguments.

pub fn update_volume(&self) -> UpdateVolume<C, M, R>

Constructs a fluent builder for the UpdateVolume operation.

See UpdateVolume for more information about the operation and its arguments.

impl<C> Client<C, AwsMiddleware, Standard>

pub fn from_conf_conn(conf: Config, conn: C) -> Self

Creates a client with the given service config and connector override.

impl Client<DynConnector, AwsMiddleware, Standard>

pub fn new(config: &Config) -> Self

Creates a new client from a shared config.

pub fn from_conf(conf: Config) -> Self

Creates a new client from the service Config.