Struct aws_sdk_opsworks::model::instance::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for Instance

Implementations

impl Builder

pub fn agent_version(self, input: impl Into<String>) -> Self

The agent version. This parameter is set to INHERIT if the instance inherits the default stack setting or to a a version number for a fixed agent version.

pub fn set_agent_version(self, input: Option<String>) -> Self

The agent version. This parameter is set to INHERIT if the instance inherits the default stack setting or to a a version number for a fixed agent version.

pub fn ami_id(self, input: impl Into<String>) -> Self

A custom AMI ID to be used to create the instance. For more information, see Instances

pub fn set_ami_id(self, input: Option<String>) -> Self

A custom AMI ID to be used to create the instance. For more information, see Instances

pub fn architecture(self, input: Architecture) -> Self

The instance architecture: "i386" or "x86_64".

pub fn set_architecture(self, input: Option<Architecture>) -> Self

The instance architecture: "i386" or "x86_64".

pub fn arn(self, input: impl Into<String>) -> Self

The instance's Amazon Resource Number (ARN).

pub fn set_arn(self, input: Option<String>) -> Self

The instance's Amazon Resource Number (ARN).

pub fn auto_scaling_type(self, input: AutoScalingType) -> Self

For load-based or time-based instances, the type.

pub fn set_auto_scaling_type(self, input: Option<AutoScalingType>) -> Self

For load-based or time-based instances, the type.

pub fn availability_zone(self, input: impl Into<String>) -> Self

The instance Availability Zone. For more information, see Regions and Endpoints.

pub fn set_availability_zone(self, input: Option<String>) -> Self

The instance Availability Zone. For more information, see Regions and Endpoints.

pub fn block_device_mappings(self, input: impl Into<BlockDeviceMapping>) -> Self

Appends an item to block_device_mappings.

To override the contents of this collection use set_block_device_mappings.

An array of BlockDeviceMapping objects that specify the instance's block device mappings.

pub fn set_block_device_mappings(
self,
input: Option<Vec<BlockDeviceMapping>>
) -> Self

An array of BlockDeviceMapping objects that specify the instance's block device mappings.

pub fn created_at(self, input: impl Into<String>) -> Self

The time that the instance was created.

pub fn set_created_at(self, input: Option<String>) -> Self

The time that the instance was created.

pub fn ebs_optimized(self, input: bool) -> Self

Whether this is an Amazon EBS-optimized instance.

pub fn set_ebs_optimized(self, input: Option<bool>) -> Self

Whether this is an Amazon EBS-optimized instance.

pub fn ec2_instance_id(self, input: impl Into<String>) -> Self

The ID of the associated Amazon EC2 instance.

pub fn set_ec2_instance_id(self, input: Option<String>) -> Self

The ID of the associated Amazon EC2 instance.

pub fn ecs_cluster_arn(self, input: impl Into<String>) -> Self

For container instances, the Amazon ECS cluster's ARN.

pub fn set_ecs_cluster_arn(self, input: Option<String>) -> Self

For container instances, the Amazon ECS cluster's ARN.

pub fn ecs_container_instance_arn(self, input: impl Into<String>) -> Self

For container instances, the instance's ARN.

pub fn set_ecs_container_instance_arn(self, input: Option<String>) -> Self

For container instances, the instance's ARN.

pub fn elastic_ip(self, input: impl Into<String>) -> Self

The instance Elastic IP address .

pub fn set_elastic_ip(self, input: Option<String>) -> Self

The instance Elastic IP address .

pub fn hostname(self, input: impl Into<String>) -> Self

The instance host name.

pub fn set_hostname(self, input: Option<String>) -> Self

The instance host name.

pub fn infrastructure_class(self, input: impl Into<String>) -> Self

For registered instances, the infrastructure class: ec2 or on-premises.

pub fn set_infrastructure_class(self, input: Option<String>) -> Self

For registered instances, the infrastructure class: ec2 or on-premises.

pub fn install_updates_on_boot(self, input: bool) -> Self

Whether to install operating system and package updates when the instance boots. The default value is true. If this value is set to false, you must then update your instances manually by using CreateDeployment to run the update_dependencies stack command or by manually running yum (Amazon Linux) or apt-get (Ubuntu) on the instances.

We strongly recommend using the default value of true, to ensure that your instances have the latest security updates.

pub fn set_install_updates_on_boot(self, input: Option<bool>) -> Self

Whether to install operating system and package updates when the instance boots. The default value is true. If this value is set to false, you must then update your instances manually by using CreateDeployment to run the update_dependencies stack command or by manually running yum (Amazon Linux) or apt-get (Ubuntu) on the instances.

We strongly recommend using the default value of true, to ensure that your instances have the latest security updates.

pub fn instance_id(self, input: impl Into<String>) -> Self

The instance ID.

pub fn set_instance_id(self, input: Option<String>) -> Self

The instance ID.

pub fn instance_profile_arn(self, input: impl Into<String>) -> Self

The ARN of the instance's IAM profile. For more information about IAM ARNs, see Using Identifiers.

pub fn set_instance_profile_arn(self, input: Option<String>) -> Self

The ARN of the instance's IAM profile. For more information about IAM ARNs, see Using Identifiers.

pub fn instance_type(self, input: impl Into<String>) -> Self

The instance type, such as t2.micro.

pub fn set_instance_type(self, input: Option<String>) -> Self

The instance type, such as t2.micro.

pub fn last_service_error_id(self, input: impl Into<String>) -> Self

The ID of the last service error. For more information, call DescribeServiceErrors.

pub fn set_last_service_error_id(self, input: Option<String>) -> Self

The ID of the last service error. For more information, call DescribeServiceErrors.

pub fn layer_ids(self, input: impl Into<String>) -> Self

Appends an item to layer_ids.

To override the contents of this collection use set_layer_ids.

An array containing the instance layer IDs.

pub fn set_layer_ids(self, input: Option<Vec<String>>) -> Self

An array containing the instance layer IDs.

pub fn os(self, input: impl Into<String>) -> Self

The instance's operating system.

pub fn set_os(self, input: Option<String>) -> Self

The instance's operating system.

pub fn platform(self, input: impl Into<String>) -> Self

The instance's platform.

pub fn set_platform(self, input: Option<String>) -> Self

The instance's platform.

pub fn private_dns(self, input: impl Into<String>) -> Self

The instance's private DNS name.

pub fn set_private_dns(self, input: Option<String>) -> Self

The instance's private DNS name.

pub fn private_ip(self, input: impl Into<String>) -> Self

The instance's private IP address.

pub fn set_private_ip(self, input: Option<String>) -> Self

The instance's private IP address.

pub fn public_dns(self, input: impl Into<String>) -> Self

The instance public DNS name.

pub fn set_public_dns(self, input: Option<String>) -> Self

The instance public DNS name.

pub fn public_ip(self, input: impl Into<String>) -> Self

The instance public IP address.

pub fn set_public_ip(self, input: Option<String>) -> Self

The instance public IP address.

pub fn registered_by(self, input: impl Into<String>) -> Self

For registered instances, who performed the registration.

pub fn set_registered_by(self, input: Option<String>) -> Self

For registered instances, who performed the registration.

pub fn reported_agent_version(self, input: impl Into<String>) -> Self

The instance's reported AWS OpsWorks Stacks agent version.

pub fn set_reported_agent_version(self, input: Option<String>) -> Self

The instance's reported AWS OpsWorks Stacks agent version.

pub fn reported_os(self, input: ReportedOs) -> Self

For registered instances, the reported operating system.

pub fn set_reported_os(self, input: Option<ReportedOs>) -> Self

For registered instances, the reported operating system.

pub fn root_device_type(self, input: RootDeviceType) -> Self

The instance's root device type. For more information, see Storage for the Root Device.

pub fn set_root_device_type(self, input: Option<RootDeviceType>) -> Self

The instance's root device type. For more information, see Storage for the Root Device.

pub fn root_device_volume_id(self, input: impl Into<String>) -> Self

The root device volume ID.

pub fn set_root_device_volume_id(self, input: Option<String>) -> Self

The root device volume ID.

pub fn security_group_ids(self, input: impl Into<String>) -> Self

Appends an item to security_group_ids.

To override the contents of this collection use set_security_group_ids.

An array containing the instance security group IDs.

pub fn set_security_group_ids(self, input: Option<Vec<String>>) -> Self

An array containing the instance security group IDs.

pub fn ssh_host_dsa_key_fingerprint(self, input: impl Into<String>) -> Self

The SSH key's Deep Security Agent (DSA) fingerprint.

pub fn set_ssh_host_dsa_key_fingerprint(self, input: Option<String>) -> Self

The SSH key's Deep Security Agent (DSA) fingerprint.

pub fn ssh_host_rsa_key_fingerprint(self, input: impl Into<String>) -> Self

The SSH key's RSA fingerprint.

pub fn set_ssh_host_rsa_key_fingerprint(self, input: Option<String>) -> Self

The SSH key's RSA fingerprint.

pub fn ssh_key_name(self, input: impl Into<String>) -> Self

The instance's Amazon EC2 key-pair name.

pub fn set_ssh_key_name(self, input: Option<String>) -> Self

The instance's Amazon EC2 key-pair name.

pub fn stack_id(self, input: impl Into<String>) -> Self

The stack ID.

pub fn set_stack_id(self, input: Option<String>) -> Self

The stack ID.

pub fn status(self, input: impl Into<String>) -> Self

The instance status:

booting

connection_lost

online

pending

rebooting

requested

running_setup

setup_failed

shutting_down

start_failed

stop_failed

stopped

stopping

terminated

terminating

pub fn set_status(self, input: Option<String>) -> Self

The instance status:

booting

connection_lost

online

pending

rebooting

requested

running_setup

setup_failed

shutting_down

start_failed

stop_failed

stopped

stopping

terminated

terminating

pub fn subnet_id(self, input: impl Into<String>) -> Self

The instance's subnet ID; applicable only if the stack is running in a VPC.

pub fn set_subnet_id(self, input: Option<String>) -> Self

The instance's subnet ID; applicable only if the stack is running in a VPC.

pub fn tenancy(self, input: impl Into<String>) -> Self

The instance's tenancy option, such as dedicated or host.

pub fn set_tenancy(self, input: Option<String>) -> Self

The instance's tenancy option, such as dedicated or host.

pub fn virtualization_type(self, input: VirtualizationType) -> Self

The instance's virtualization type: paravirtual or hvm.

pub fn set_virtualization_type(self, input: Option<VirtualizationType>) -> Self

The instance's virtualization type: paravirtual or hvm.

pub fn build(self) -> Instance

Consumes the builder and constructs a Instance